reorganize project, removing client/server organization after realizing CPU utilization is all due to the client, so having a server component wouldnt make much sense
