--- cmd/prompt.go
@@ -1,35 +1,0 @@
-/*
-Copyright © 2025 Greg Griffin <greg.griffin2@gmail.com>
-*/
-package cmd
-
-import (
-	"fmt"
-
-	"github.com/spf13/cobra"
-)
-
-// promptCmd represents the prompt command
-var promptCmd = &cobra.Command{
-	Use:   "prompt [model to prompt]",
-	Short: "Begin a prompt session with a model using an existing server",
-	Long:  `Assuming a server has been started with 'llm serve', begin a prompt session with a specified model`,
-	Args:  cobra.MinimumNArgs(1),
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("prompt called")
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(promptCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// promptCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// promptCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
