+++ cmd/run.go
@@ -4,8 +4,8 @@ Copyright © 2025 Greg Griffin <greg.griffin2@gmail.com>
 package cmd
 
 import (
-	"github.com/gregriff/gpt-cli-go/client/repl"
 	"github.com/gregriff/gpt-cli-go/models/anthropic"
+	"github.com/gregriff/gpt-cli-go/repl"
 	"github.com/spf13/cobra"
 )
 
@@ -15,7 +15,7 @@ var modelName string
 var runCmd = &cobra.Command{
 	Use:   "run [model to prompt]",
 	Short: "Create a new prompt session with a model",
-	Long:  `Begin a prompt session with a specified model, creating a new server for convenience. Server will be shutdown on exit`,
+	Long:  `Begin a prompt session with a specified model.`,
 	// Args:  cobra.MinimumNArgs(1),
 	PreRunE: func(cmd *cobra.Command, args []string) error {
 		return anthropic.ValidateModelName(modelName)
