--- cmd/serve.go
@@ -1,35 +1,0 @@
-/*
-Copyright © 2025 Greg Griffin <greg.griffin2@gmail.com>
-*/
-package cmd
-
-import (
-	"fmt"
-
-	"github.com/spf13/cobra"
-)
-
-// serveCmd represents the serve command
-var serveCmd = &cobra.Command{
-	Use:   "serve [options]",
-	Short: "Start a server that handles any number of prompt sessions",
-	Long:  `Starts a local server to handle prompt sessions. To begin a prompt session, run the 'llm prompt' command in a seperate shell`,
-	Args:  cobra.MinimumNArgs(1),
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("serve called")
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(serveCmd)
-
-	// Here you will define your flags and configuration settings.
-
-	// Cobra supports Persistent Flags which will work for this command
-	// and all subcommands, e.g.:
-	// serveCmd.PersistentFlags().String("foo", "", "A help for foo")
-
-	// Cobra supports local flags which will only run when this command
-	// is called directly, e.g.:
-	// serveCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
